Big rewrite of RobustTcpTemplate class.
Main emphasis was placed on potential problems after the connection is lost.
o Transaction has timeout and may expire
o After connection is lost, all of the outstanding transactions are released.
o If connection is not estabilished, new requests are refused.
o It is possible to stop the request / response loop.
o Better identification of the request / response thread for debugging.